==> scripts/notebooks/10_prepareDatasets.ipynb <==
# Prepare dataset for analysis (GSE87571)

- AGE dataset (GSE87571) --> downloaded from methylnet github
- parse only CpGs from a specified chromosome

import numpy as np
import torch
import pickle
import pandas as pd
import os
from data.prepareData import prepareDataLoader_fromPkl

PATH_DATA = "/data/scratch/skatz/PROJECTS/methylnet/1_healthyVAE/data"
PATH_manifest = "/data/scratch/skatz/PROJECTS/methylnet"
DATASET = "GSE87571"

DATASET

### Load Manifest
df_manifest = pd.read_csv(f"{PATH_manifest}/humanmethylation450_15017482_v1-2.csv", skiprows=6, header=1, index_col=0, low_memory=False)

'''
Per Chromosome parsing
'''
for ele in range(1,2):
    print(f"\n\n{ele}")
    chr = str(ele)
    ## select relevant chromosome to parse
    probes_chr = df_manifest[df_manifest["CHR"] == chr].index.tolist()
    print(len(probes_chr))
    
    ### Find overlap between probes form Chr and what's left after preprocessing
    ds = pickle.load(open(f"{PATH_DATA}/{DATASET}/train_val_test_sets/train_methyl_array.pkl", "rb"))
    cpgs = ds["beta"].columns[ds["beta"].columns.isin(probes_chr)]
    print(len(cpgs))

### Load Dataset
for file in ["train_methyl_array.pkl", "test_methyl_array.pkl", "val_methyl_array.pkl"]:
    print(file)
    ds = pickle.load(open(f"{PATH_DATA}/{DATASET}/train_val_test_sets/{file}", "rb"))

    ### save only those probes
    ds["beta"] = ds["beta"].loc[:,cpgs]
    print(f"Size: {ds['beta'].shape}")
    #
    pickle.dump(ds, open(f"{PATH_DATA}/{DATASET}/train_val_test_sets/chr{chr}_{file}", "wb"))
