==> scripts/notebooks/30_optimisation_fineGridSearch.ipynb <==
# Design fine-grained model search

- Read in best (latent size) model from coarse grained
- Define latent sizes in close vicinity (+ / -) 
(- Same `lr` and `dropout` scan as in `coarse`)

import os
import pandas as pd
import numpy as np
import json
import pickle 
import torch
import itertools
import time

## Check `lr` and `dropout` of best models for all chromosomes
- Questions: can we reduce grid search space for those parameters? 

dict_overviewBestModelParams = dict(lr=[], dropr=[])
### Parse hyperparameters of best models
for i in range(1,23):
    CHR = f"chr{i}"
    PATH_results = f"logs/optimisation/{CHR}/coarse"
    with open(f"{PATH_results}/best_model_coarseOptimization.json", "r") as f: dict_bestModel = json.load(f)
    dict_overviewBestModelParams["lr"].append(dict_bestModel["lr"])
    dict_overviewBestModelParams["dropr"].append(dict_bestModel["dropr"])
### Print hp count
print(pd.Series(dict_overviewBestModelParams["lr"]).value_counts())
print(pd.Series(dict_overviewBestModelParams["dropr"]).value_counts())  

# Design fine-search grid

for CHR in ["chr6", "chr8"]:#[f"chr{i}" for i in range(1,15)]:
    PATH_results = f"logs/optimisation/{CHR}/fine"
    os.makedirs(PATH_results, exist_ok=True)
    param_grid = dict()

    ### Step 1: read in old model
    ## a. original parameters
    with open(f"logs/optimisation/{CHR}/coarse/param_grid_coarse.json", "r") as f: dict_oldGrid = json.load(f)
    ## b. best HP model
    with open(f"logs/optimisation/{CHR}/coarse/best_model_coarseOptimization.json", "r") as f: dict_bestModel = json.load(f)

    ### Step 2: design hidden layers --> same as in coarse grid!
    param_grid["hidden_layer_encoder_topology"] = dict_oldGrid["hidden_layer_encoder_topology"]

    ### Step 3: design latSizes
    ## Get best model and its neighboring latSizes 
    idx_best = dict_oldGrid["latentSize_coarse"].index(int(dict_bestModel["latSize"]))
    bestLatNeighbors = [dict_oldGrid["latentSize_coarse"][idx_best+i] for i in range(-1,2)]
    ## Define points in between neighbors
    lowerLatSizes = np.linspace(bestLatNeighbors[2],bestLatNeighbors[1],5)[1:].round().tolist()
    upperLatSizes = np.linspace(bestLatNeighbors[1],bestLatNeighbors[0],5)[1:-1].round().tolist()
    param_grid["latentSize_fine"] = list(map(int,lowerLatSizes + upperLatSizes))

    ### Step 4: decide on lr and dropout rates to try
    ### reduce searchable space according to experiment above
    lr_scan = [1e-4, 5e-4]#, 1e-3] 
    param_grid["lr"] = lr_scan
    dropout_scan = [0.1,0.3]
    param_grid["dropout"] = dropout_scan

    ### Save parameter grid in file for later documentation
    with open(f"{PATH_results}/param_grid_fine.json", "w") as f: f.write(json.dumps(param_grid, indent="\t"))

for CHR in ["chr6", "chr8"]:#[f"chr{i}" for i in range(1,15)]:
    PATH_results = f"logs/optimisation/{CHR}/fine"
    ### Step 5: generate combinations of parameters and replace in file; standaline script?
    ### Load parameter grid
    with open(f"{PATH_results}/param_grid_fine.json", "r") as f: param_grid_comb=json.loads(f.read())

    ### Generate combinations; remove 'hidden_layer_encoder_topology' before, as it should be fixed
    param_grid = param_grid_comb.copy()
    param_grid_comb.pop("hidden_layer_encoder_topology", None)
    combs_coarse = [dict(zip(param_grid_comb.keys(), values)) for values in itertools.product(*param_grid_comb.values())]
    print(f"Number of combinations: {len(combs_coarse)}")

    ### Generate submit.sh with combinations
    os.makedirs(f"{PATH_results}/submit", exist_ok=True)
    with open("submit_template.sh", "r") as f: template=f.read()
    for i in range(len(combs_coarse)):                          
        latSize = str(combs_coarse[i]["latentSize_fine"])
        lr = str(combs_coarse[i]["lr"])
        dropr = str(combs_coarse[i]["dropout"])

        ### Generate run name - combination of parameter settings
        fileName = f"latSize_{latSize}_lr_{lr}_dropr_{dropr}"
        ### Replace in template file
        template_updated = template.replace("$PATH", str(PATH_results+"/"+fileName)) \
                                   .replace("$CHR",  str(CHR)) \
                                   .replace("$HIDDEN_1", str(param_grid["hidden_layer_encoder_topology"][0])) \
                                   .replace("$HIDDEN_2", str(param_grid["hidden_layer_encoder_topology"][1])) \
                                   .replace("$LATSIZE", latSize) \
                                   .replace("$LR", lr) \
                                   .replace("$DROPR", dropr)
        with open(f"{PATH_results}/submit/{fileName}.sh", "w") as f: f.write(template_updated)
        print(f"Wrote file \t{fileName}")

# Overview all fine grained jobs

PATH_data = "/data/scratch/skatz/PROJECTS/methylnet/1_healthyVAE/data/GSE87571/train_val_test_sets/"

df_overview = pd.DataFrame(index=[f"chr{ele}" for ele in range(1,23)], columns=["num_cpgs"]+list(param_grid.keys()))

for CHR in [f"chr{ele}" for ele in range(1,23)]:
    ### Get number of input CpGs
    with open(os.path.join(PATH_data, f"{CHR}_train_methyl_array.pkl"), "rb") as f: train_dataset = pickle.load(f) #
    num_cpgs = train_dataset["beta"].shape[1]
    
    ### Parse parameter grid
    PATH_results = f"logs/optimisation/{CHR}/fine"
    with open(f"{PATH_results}/param_grid_fine.json", "r") as f: param = json.load(f)
        
    ### add to output dataframe
    df_overview.loc[CHR,"num_cpgs"] = num_cpgs
    df_overview.loc[CHR,"hidden_layer_encoder_topology"] = param["hidden_layer_encoder_topology"]
    df_overview.loc[CHR,"latentSize_fine"] = param["latentSize_fine"]
    df_overview.loc[CHR,"lr"] = param["lr"]
    df_overview.loc[CHR,"dropout"] = param["dropout"]
    


df_overview
