==> scripts/notebooks/32_optimisation_finalModels_smallestBottleneck.ipynb <==
# Run all chromosomes without Bottleneck - maximum recon?

- Read in best model from fine-grained
- Rerun

import os
import pandas as pd
import numpy as np
import json
import pickle 
import torch

# Find best models

PATH_data = "/data/scratch/skatz/PROJECTS/methylnet/1_healthyVAE/data/GSE87571/train_val_test_sets/"

for CHR in [f"chr{i}" for i in range(1,22)]:
    PATH_results = f"logs/finalModels/{CHR}/smallestBottleneck"
    os.makedirs(PATH_results, exist_ok=True)
    param_grid = dict()

    ### Step 1: read in old model
    ## a. original parameters
    with open(f"logs/optimisation/{CHR}/fine/param_grid_fine.json", "r") as f: dict_oldGrid = json.load(f)
    ## b. best HP model
    with open(f"logs/optimisation/{CHR}/fine/best_model_fineOptimization.json", "r") as f: dict_bestModel = json.load(f)
        
    ### Step 2: design hidden layers --> same as in coarse grid!
    param_grid["hidden_layer_encoder_topology"] = dict_oldGrid["hidden_layer_encoder_topology"]

    ### Step 3: design latSizes
    with open(os.path.join(PATH_data, f"{CHR}_train_methyl_array.pkl"), "rb") as f: train_dataset = pickle.load(f) #
    num_cpgs = train_dataset["beta"].shape[1]
    param_grid["latentSize"] = 1
    ### Step 4: get lr and dropout
    param_grid["lr"] = dict_bestModel["lr"]
    param_grid["dropout"] = dict_bestModel["dropr"]

    ### Save parameter grid in file for later documentation
    with open(f"{PATH_results}/param_grid.json", "w") as f: f.write(json.dumps(param_grid, indent="\t"))

for CHR in [f"chr{i}" for i in range(1,22)]:
    PATH_results = f"logs/finalModels/{CHR}/smallestBottleneck"
    ### Load parameter grid
    with open(f"{PATH_results}/param_grid.json", "r") as f: param_grid_comb=json.loads(f.read())

    ### Generate submit.sh with combinations
    os.makedirs(f"{PATH_results}/submit", exist_ok=True)
    with open("submit_template.sh", "r") as f: template=f.read()      
    latSize = str(param_grid_comb["latentSize"])
    lr = str(param_grid_comb["lr"])
    dropr = str(param_grid_comb["dropout"])

    ### Generate run name - combination of parameter settings
    fileName = f"latSize_{latSize}"
    ### Replace in template file
    template_updated = template.replace("$PATH", str(PATH_results+"/"+fileName)) \
                                   .replace("$CHR",  str(CHR)) \
                                   .replace("$HIDDEN_1", str(param_grid_comb["hidden_layer_encoder_topology"][0])) \
                                   .replace("$HIDDEN_2", str(param_grid_comb["hidden_layer_encoder_topology"][1])) \
                                   .replace("$LATSIZE", latSize) \
                                   .replace("$LR", lr) \
                                   .replace("$DROPR", dropr)
    with open(f"{PATH_results}/submit/{fileName}.sh", "w") as f: f.write(template_updated)
    print(f"Wrote file \t{fileName}")

# Analysis

from scipy.stats import pearsonr
def calc_r2(model, data_tensor):    
    model.eval()
    orig = data_tensor.cpu().detach().numpy()
    recon = model(data_tensor)
    # check if VAE or AE was used
    if isinstance(recon, tuple):
        recon = recon[0].cpu().detach().numpy()
    else:
        recon = recon.detach().numpy()
    r2 = []
    for i in range(recon.shape[1]):
        r2.append(pearsonr(orig[:,i], recon[:,i])[0])
    return np.array(r2).mean().round(3), np.array(r2).std().round(3)

PATH_data = "/data/scratch/skatz/PROJECTS/methylnet/1_healthyVAE/data/GSE87571/train_val_test_sets/"

all_mean = []
all_std = []
for CHR in [f"chr{i}" for i in range(1,23)]:
    with open(os.path.join(PATH_data, f"{CHR}_test_methyl_array.pkl"), "rb") as f: test_dataset = pickle.load(f) #
    test_tensor = torch.tensor(test_dataset["beta"].values, dtype=torch.float32)
    
    PATH_results = f"logs/finalModels/{CHR}/smallestBottleneck"
    with open(f"{PATH_results}/param_grid.json", "r") as f: dict_bestModel = json.load(f)
    latSize = dict_bestModel["latentSize"]
    name = f"latSize_{latSize}"
    path = f"{PATH_results}/{name}"
    model = torch.load(f"{path}/checkpoint/trainedModel.pth", map_location=torch.device('cpu'))
    
    mean, std = calc_r2(model, test_tensor)
    print(CHR, mean, std)
    all_mean.append(mean)
    all_std.append(std)

print(np.nanmean(np.array(all_mean)))
print(np.nanmean(np.array(all_std)))
