==> scripts/notebooks/99_cleanCluster.ipynb <==
# Clean cluster

import os 
import sys
import json

delete_sure = True

if delete_sure:
    for CHR in [f"chr{i}" for i in range(1,23)]: # 23
        ### Get best model
        PATH_results = f"logs/optimisation/{CHR}/fine"
        print(PATH_results)
        with open(f"{PATH_results}/best_model_fineOptimization.json", "r") as f: dict_bestModel = json.load(f)
        latSize = dict_bestModel["latSize"]
        lr = dict_bestModel["lr"]
        dropr = dict_bestModel["dropr"]    
        name = f"latSize_{latSize}_lr_{lr}_dropr_{dropr}"
        print(name)
    
        ### Deleete all .pth but best model
        for folder in os.listdir(PATH_results):
            if folder.startswith("latSize_"):
                if folder.startswith(name):
                    print(f"Spared best model: {name}")
                else:
                    try:
                        os.remove(f"{PATH_results}/{folder}/checkpoint/trainedModel.pth")
                    except FileNotFoundError:
                        continue
